ObsidynOsSOurceCode/requirements.txt
argon2-cffi==23.1.0
cryptography==41.0.7
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.3.0
